+++ engine/code/game/g_local.h
@@ -429,6 +429,8 @@ qboolean	G_SpawnFloat( const char *key, const char *defaultString, float *out );
 qboolean	G_SpawnInt( const char *key, const char *defaultString, int *out );
 qboolean	G_SpawnVector( const char *key, const char *defaultString, float *out );
 void		G_SpawnEntitiesFromString( void );
+void		G_SpawnDynamicPickups( void );
+void		G_RegisterDynamicPickups( void );
 char *G_NewString( const char *string );
 
 //