+++ RELEASE_NOTES.md
@@ -2,6 +2,10 @@
 
 ## Current Release
 
+### New Features:
+
+1.  Add the ability to spawn pickups dynamically at arbitrary locations.
+
 ### Minor Improvements:
 
 1.  Moved .map files into assets/maps/src and .bsp files into assets/maps/built.